+++ src/test/java/racing/CarTest.java
@@ -3,6 +3,7 @@ package racing;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.Test;
+import racing.domain.Car;
 
 public class CarTest {
 
@@ -19,8 +20,5 @@ public class CarTest {
     car.driveOrStop(5);
     assertThat(car.getPosition()).isEqualTo(1);
   }
-
-
-
 }
