+++ src/test/java/racing/RacingTest.java
@@ -3,6 +3,7 @@ package racing;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.Test;
+import racing.domain.Racing;
 
 class RacingTest {
